chore: streamline pipeline structure
